Authorization Added in All Cart Views

+++ ps_project/ps_app/views/cart_view.py
@@ -7,6 +7,10 @@ from ..models import Cart, Product, CartItem
 def add_to_cart_view(request, product_id):
     product = Product.objects.get(id=product_id)
     
+    if request.user.is_staff:
+        messages.error(request, "Staff members cannot add products to the cart.")
+        return redirect('/dashboard/admin/')
+    
     if product.stock <= 0:
         messages.warning(request, "This product is out of stock.")
         return redirect('/products')
@@ -18,6 +22,10 @@ def add_to_cart_view(request, product_id):
 
 @login_required
 def cart_view(request):
+    if request.user.is_staff:
+        messages.error(request, "Staff members cannot have a cart.")
+        return redirect('/dashboard/admin/')
+    
     cart, created = Cart.objects.get_or_create(customer=request.user)
     cart_items = cart.items.all()
     total_price = sum(item.total_price for item in cart_items)
@@ -27,6 +35,10 @@ def cart_view(request):
 def increase_cart_item_quantity(request, product_id):
     cart = Cart.objects.get(customer=request.user)
     cart_item = CartItem.objects.get(cart=cart, product_id=product_id)
+    
+    if request.user.is_staff:
+        messages.error(request, "Staff members cannot modify cart items.")
+        return redirect('/dashboard/admin/')
     
     if cart_item.product.stock <= cart_item.quantity:
         messages.warning(request, "Cannot increase quantity. Stock limit reached.")
@@ -38,6 +50,10 @@ def increase_cart_item_quantity(request, product_id):
 
 @login_required
 def decrease_cart_item_quantity(request, product_id):
+    if request.user.is_staff:
+        messages.error(request, "Staff members cannot modify cart items.")
+        return redirect('/dashboard/admin/')
+
     cart = Cart.objects.get(customer=request.user)
     cart_item = CartItem.objects.get(cart=cart, product_id=product_id)
     if cart_item.quantity > 1:
@@ -49,6 +65,10 @@ def decrease_cart_item_quantity(request, product_id):
 
 @login_required
 def remove_cart_item(request, product_id):
+    if request.user.is_staff:
+        messages.error(request, "Staff members cannot modify cart items.")
+        return redirect('/dashboard/admin/')
+    
     cart = Cart.objects.get(customer=request.user)
     cart_item = CartItem.objects.get(cart=cart, product_id=product_id)
     cart_item.delete()
